refactor(tenant-transactions): email services for booking confirmations, rejections, and reminders
