+++ src/routers/tenant-transactions/confirm-order/send.confirmation.service.ts
@@ -4,7 +4,7 @@ import Handlebars from 'handlebars';
 import transporter from '../../../lib/config/nodemailer.transporter';
 import database from '../../../lib/config/prisma.client';
 
-export class BookingEmailService {
+export class SendConfirmationService {
   private static async sendEmail(
     to: string,
     subject: string,
@@ -121,130 +121,6 @@ export class BookingEmailService {
       'confirmed.booking.html',
       templateData,
       attachment,
-    );
-  }
-
-  static async sendBookingRejectionEmail(bookingId: number): Promise<void> {
-    const booking = await database.booking.findUnique({
-      where: { id: bookingId },
-      include: {
-        property: true,
-        room: true,
-      },
-    });
-
-    if (!booking) {
-      throw new Error('Booking not found');
-    }
-
-    const checkInDate = new Date(booking.check_in_date).toLocaleDateString(
-      'en-US',
-      {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-      },
-    );
-
-    const checkOutDate = new Date(booking.check_out_date).toLocaleDateString(
-      'en-US',
-      {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-      },
-    );
-
-    const paymentDeadline = new Date(
-      booking.payment_deadline,
-    ).toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-
-    const templateData = {
-      order_id: booking.uid || `BK${booking.id.toString().padStart(10, '0')}`,
-      property_title: booking.property.title,
-      property_address: `${booking.property.address}, ${booking.property.city}, ${booking.property.country} ${booking.property.postal_code}`,
-      room_name: booking.room.name,
-      check_in_date: checkInDate,
-      check_out_date: checkOutDate,
-      fullname: booking.fullname,
-      email: booking.email,
-      phone_number: booking.phone_number,
-      total_price: new Intl.NumberFormat('id-ID', {
-        style: 'currency',
-        currency: 'IDR',
-        minimumFractionDigits: 0,
-      }).format(Number(booking.total_price)),
-      payment_deadline: paymentDeadline,
-      current_year: new Date().getFullYear(),
-    };
-
-    await this.sendEmail(
-      booking.email,
-      `Payment Proof Rejected - ${templateData.order_id}`,
-      'rejected.booking.html',
-      templateData,
-    );
-  }
-
-  static async sendBookingReminderEmail(bookingId: number): Promise<void> {
-    const booking = await database.booking.findUnique({
-      where: { id: bookingId },
-      include: {
-        property: true,
-        room: true,
-      },
-    });
-
-    if (!booking) {
-      throw new Error('Booking not found');
-    }
-
-    const checkInDate = new Date(booking.check_in_date).toLocaleDateString(
-      'en-US',
-      {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-      },
-    );
-
-    const checkOutDate = new Date(booking.check_out_date).toLocaleDateString(
-      'en-US',
-      {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-      },
-    );
-
-    const templateData = {
-      uid: booking.uid || `BK${booking.id.toString().padStart(10, '0')}`,
-      property_name: booking.property.title,
-      room_name: booking.room.name,
-      check_in_date: checkInDate,
-      check_out_date: checkOutDate,
-      fullname: booking.fullname,
-      total_price: new Intl.NumberFormat('id-ID', {
-        style: 'currency',
-        currency: 'IDR',
-        minimumFractionDigits: 0,
-      }).format(Number(booking.total_price)),
-      booking_detail_url: `https://staysia.com/booking/${booking.uid || `BK${booking.id.toString().padStart(10, '0')}`}`,
-      check_in_time: '14:00', // Assuming default check-in time, can be made configurable
-      current_year: new Date().getFullYear(),
-    };
-
-    await this.sendEmail(
-      booking.email,
-      `Reminder: Your stay begins tomorrow - ${templateData.uid}`,
-      'order.reminder.html',
-      templateData,
     );
   }
 }
